+++ lib/core/themes/styles.dart
@@ -2,7 +2,7 @@ import 'package:diva/core/themes/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import 'FontWeightHelper.dart';
+import 'Font_Weight_Helper.dart';
 
 class TextStyles {
   static TextStyle font14GreyRegular = TextStyle(
